Sideload analyzer feedback for SolutionChannel
