+++ app/channels/solution_channel.rb
@@ -10,6 +10,6 @@ class SolutionChannel < ApplicationCable::Channel
   def self.broadcast!(solution)
     broadcast_to solution,
       solution: SerializeSolution.(solution),
-      iterations: solution.iterations.map { |iteration| SerializeIteration.(iteration) }
+      iterations: solution.iterations.map { |iteration| SerializeIteration.(iteration, sideload: [:automated_feedback]) }
   end
 end
